Require matching types in precision checks

+++ test/unit/doc/relative.cpp
@@ -13,8 +13,8 @@
 TTS_CASE("Relative distance")
 {
   TTS_RELATIVE_EQUAL(42.f, 42.f, 0);
-  TTS_RELATIVE_EQUAL('A', 80LL, 18.75);
-  TTS_RELATIVE_EQUAL(1., 2.f, 100.);
+  TTS_RELATIVE_EQUAL(static_cast<long long>('A'), 80LL, 18.75);
+  TTS_RELATIVE_EQUAL(1., 2., 100.);
   TTS_RELATIVE_EQUAL(1, 10, 900., REQUIRED);
 };
 //! [snippet]
